Refactor: TransType 추가, 금액 BigDecimal -> Integer로 변경

+++ src/main/java/com/ddiring/BackEnd_Escrow/dto/request/SaveRecordRequest.java
@@ -15,5 +15,5 @@ public class SaveRecordRequest {
     private String userSeq;
     private Integer transSeq;
     private Integer transType;
-    private BigDecimal amount;
+    private Integer amount;
 }
